A generic function for cleaning data ready for analysis

A generic function for cleaning data ready for analysis

Usage

dataComb(CFM, DC, id = NULL, trt = NULL, cfmOb = FALSE)

Arguments

CFM

a model object supplied to pscfit

DC

a dataset including covariates to match the CFM

id

to specify which observations in the data cohort should be evaluated. Defualts to 'NULL' i.e all observations

trt

used to specify multiple treatment effects. Defaults to NULL

cfmOb

used to specify if a CFM object is supplies as the counter factual model

Value

datComb returns a list containing objects which detial the components of both the Counter Factual Model (CFM) and the Data Cohort (DC) the required exported components of the model and a cleaned data cohort.

Examples

bin.mod <- psc::bin.mod
#> Error: 'bin.mod' is not an exported object from 'namespace:psc'
data <- psc::data
#> Error: 'data' is not an exported object from 'namespace:psc'
dc <- dataComb(bin.mod,data)
#> Error: object 'bin.mod' not found