Fucntion for estimating initial parameter values

Fucntion for estimating initial parameter values

Usage

# S3 method for class 'glm'
initParm(CFM, DC_clean, trt = trt)

Arguments

CFM

A counter-factual model

DC_clean

a cleaned dataset obsect obtained using dataComb.flexsurvreg

trt

An optional additional vector denoting treatment allocations for multiple treatment comparisons. Defaults to 'NULL'

Value

an 'optim' output giving the parameter values to be supplied as a starting value for the mcmc routine.

Details

This function takes the liklihood for a 'flexsurvreg' model and uses 'optim' to fit the likelihood.